moved functions to module
